Added CPU ID feature detection of AVX 10.x and APX

+++ kernels/common/device.cpp
@@ -188,7 +188,7 @@ namespace embree
 
   void Device::print()
   {
-    const int cpu_features = getCPUFeatures();
+    const int64_t cpu_features = getCPUFeatures();
     std::cout << std::endl;
     std::cout << "Embree Ray Tracing Kernels " << RTC_VERSION_STRING << " (" << RTC_HASH << ")" << std::endl;
     std::cout << "  Compiler  : " << getCompilerName() << std::endl;
