+++ kernels/common/state.h
@@ -138,8 +138,8 @@ namespace embree
     size_t numUserThreads;                 //!< number of user provided threads to use in builders
     bool set_affinity;                     //!< sets affinity for worker threads
     bool start_threads;                    //!< true when threads should be started at device creation time
-    int enabled_cpu_features;              //!< CPU ISA features to use
-    int enabled_builder_cpu_features;      //!< CPU ISA features to use for builders only
+    int64_t enabled_cpu_features;          //!< CPU ISA features to use
+    int64_t enabled_builder_cpu_features;  //!< CPU ISA features to use for builders only
     enum FREQUENCY_LEVEL {
       FREQUENCY_SIMD128,
       FREQUENCY_SIMD256,
